requirements.txt
pillow>=9.0.0
pytesseract>=0.3.8
pymupdf>=1.19.0